Crée un composant dédiée au tableau des endpoints utiles
