+++ src/custom/simplifions/components/SimplifionsReco.vue
@@ -76,87 +76,12 @@
             </a>
           </div>
 
-          <div v-if="usefulDataApiFourniesParLaSolution === undefined">
-            Chargement des données en cours...
-          </div>
-          <div v-else-if="sortedUsefulEndpoints?.length">
-            <p class="fr-mt-4w"><b><span title="Un endpoint est une sous-partie d'une API, un point d'accès spécifique à l'intérieur de l'API." style="text-decoration: underline dotted; text-underline-offset: 3px;">Endpoints</span> de l'API utiles pour ce cas d'usage :</b></p>
-            <div class="fr-table fr-table--multiline fr-table--no-caption fr-table--sm">
-              <div class="fr-table__header">
-                <div class="fr-search-bar" role="search">
-                  <label class="fr-label" :for="searchId">Filtrer les endpoints</label>
-                  <input
-                    :id="searchId"
-                    v-model="searchQuery"
-                    class="fr-input"
-                    type="search"
-                    placeholder="Rechercher les endpoints utiles par mots-clés..."
-                  />
-                  <button type="button" class="fr-btn">Filtrer</button>
-                </div>
-              </div>
-              <div ref="tableWrapperRef" class="fr-table__wrapper">
-                <div class="fr-table__container">
-                  <div class="fr-table__content">
-                    <table>
-                      <thead>
-                        <tr class="">
-                          <th class="fr-col--md fr-py-2w" scope="col">Endpoints utiles de l'API</th>
-                          <th class="fr-col--lg fr-py-2w" scope="col">Description de l'utilité pour le cas d'usage «&nbsp;{{ recommandation.Nom_complet_du_cas_d_usage }}&nbsp;»</th>
-                        </tr>
-                      </thead>
-                      <tbody ref="tableBodyRef">
-                        <tr v-if="!filteredEndpoints.length">
-                          <td colspan="2"><i>Aucun endpoint ne correspond à votre recherche.</i></td>
-                        </tr>
-                        <template v-else>
-                          <tr
-                            v-for="record in displayedEndpoints"
-                            :key="record.fields.UID_datagouv"
-                            class="test__api-or-dataset-utile"
-                          >
-                            <td>
-                              <b>{{ record.fields.Nom }}</b><br />
-                              <a
-                                :href="`https://www.data.gouv.fr/fr/${datagouvUrlType(record.fields)}/${record.fields.UID_datagouv}`"
-                                target="_blank"
-                                class="fr-link fr-link--xs"
-                              >Documentation</a>
-                            </td>
-                            <!-- eslint-disable vue/no-v-html -->
-                            <td
-                              v-if="customDescriptions[record.id]?.En_quoi_cette_API_ou_dataset_est_utile_pour_ce_cas_d_usage"
-                              class="fr-text--sm fr-py-1w test__api-or-dataset-utile-description"
-                              v-html="fromMarkdown(customDescriptions[record.id].En_quoi_cette_API_ou_dataset_est_utile_pour_ce_cas_d_usage).html"
-                            ></td>
-                            <!-- eslint-enable vue/no-v-html -->
-                            <td v-else>
-                              <i class="fr-text--xs">Cet endpoint est identifié comme utile pour ce cas d'usage.</i>
-                            </td>
-                          </tr>
-                        </template>
-                      </tbody>
-                    </table>
-                  </div>
-                </div>
-              </div>
-              <div v-if="paginationPages.length > 1" class="fr-table__footer">
-                <div class="fr-table__footer--start">
-                  <p class="fr-table__detail">{{ filteredEndpoints.length }} endpoint{{ filteredEndpoints.length > 1 ? 's' : '' }}</p>
-                </div>
-                <div class="fr-table__footer--middle">
-                  <DsfrPagination
-                    :pages="paginationPages"
-                    :current-page="currentPage"
-                    @update:current-page="handlePageChange"
-                  />
-                </div>
-              </div>
-            </div>
-          </div>
-
-
-
+          <SimplifionsRecoUsefulEndpointsTable
+            :endpoints="usefulDataApiFourniesParLaSolution"
+            :custom-descriptions="customDescriptions"
+            :case-usage-name="recommandation.Nom_complet_du_cas_d_usage"
+            :active="activeAccordion === API_ACCORDION_INDEX"
+          />
         </template>
 
         <template v-else>
@@ -257,6 +182,7 @@ import type {
 import TopicsAPI from '../simplifionsTopicsApi'
 import SimplifionsDataApi from './SimplifionsDataApi.vue'
 import SimplifionsRecoIntegratingSolutionsAccordion from './SimplifionsRecoIntegratingSolutionsAccordion.vue'
+import SimplifionsRecoUsefulEndpointsTable from './SimplifionsRecoUsefulEndpointsTable.vue'
 
 const props = defineProps<{
   recommandation: Recommandation
@@ -269,114 +195,6 @@ const isSolution = !!recommandation.Solution_recommandee
 const topicSlug = ref<string | undefined>(undefined)
 const usefulDataApiFourniesParLaSolution = ref<ApiOrDatasetRecord[] | undefined>(undefined)
 const customDescriptions = ref<Record<number, ApiOrDatasetUtiles>>({})
-const searchId = useId()
-const searchQuery = ref('')
-const currentPage = ref(0)
-const tableWrapperRef = ref<HTMLElement | null>(null)
-const tableBodyRef = ref<HTMLElement | null>(null)
-const pageBreaks = ref<number[]>([0])
-const isMeasuring = ref(false)
-const usableHeight = ref(0)
-
-const updateRowHeight = () => {
-  if (!tableWrapperRef.value || !usableHeight.value) return
-  const rowsOnPage = paginatedEndpoints.value.length
-  if (!rowsOnPage) {
-    tableWrapperRef.value.style.removeProperty('--row-height')
-    return
-  }
-  const rowH = usableHeight.value / rowsOnPage
-  tableWrapperRef.value.style.setProperty('--row-height', `${rowH}px`)
-}
-
-const handlePageChange = async (page: number) => {
-  currentPage.value = page
-  await nextTick()
-  updateRowHeight()
-}
-
-const recomputePageBreaks = async () => {
-  const wrapper = tableWrapperRef.value
-  if (!wrapper || !filteredEndpoints.value.length) return
-
-  isMeasuring.value = true
-  await nextTick()
-
-  // Remove height constraint to measure natural row heights
-  wrapper.style.removeProperty('--row-height')
-
-  const theadHeight = tableBodyRef.value?.closest('table')?.querySelector('thead')?.offsetHeight ?? 0
-  // clientHeight excludes wrapper borders (1px × 2 = 2px)
-  usableHeight.value = wrapper.clientHeight - theadHeight
-
-  const rows = [...(tableBodyRef.value?.querySelectorAll('tr') ?? [])]
-  const breaks = [0]
-  let pageStart = 0
-  let maxH = 0
-
-  for (let i = 0; i < rows.length; i++) {
-    const h = (rows[i] as HTMLElement).offsetHeight
-    const rowsAfterAdd = i - pageStart + 1
-    const newMax = Math.max(maxH, h)
-
-    // Invariant: rowsAfterAdd × max(heights) ≤ usableHeight
-    // Ensures no row overflows after uniform stretching to usableHeight/n
-    if (rowsAfterAdd > 1 && rowsAfterAdd * newMax > usableHeight.value) {
-      breaks.push(i)
-      pageStart = i
-      maxH = h
-    } else {
-      maxH = newMax
-    }
-  }
-
-  pageBreaks.value = breaks
-  isMeasuring.value = false
-  await nextTick()
-  updateRowHeight()
-}
-
-let resizeTimer: ReturnType<typeof setTimeout> | undefined
-const debouncedRecomputePageBreaks = () => {
-  clearTimeout(resizeTimer)
-  resizeTimer = setTimeout(recomputePageBreaks, 150)
-}
-
-onMounted(() => window.addEventListener('resize', debouncedRecomputePageBreaks))
-onUnmounted(() => {
-  window.removeEventListener('resize', debouncedRecomputePageBreaks)
-  clearTimeout(resizeTimer)
-})
-
-const filteredEndpoints = computed(() => {
-  const q = searchQuery.value.trim().toLowerCase()
-  if (!q) return sortedUsefulEndpoints.value ?? []
-  return (sortedUsefulEndpoints.value ?? []).filter((record) => {
-    const description = customDescriptions.value[record.id]?.En_quoi_cette_API_ou_dataset_est_utile_pour_ce_cas_d_usage ?? ''
-    return (
-      record.fields.Nom.toLowerCase().includes(q) ||
-      description.toLowerCase().includes(q)
-    )
-  })
-})
-
-const paginatedEndpoints = computed(() => {
-  const start = pageBreaks.value[currentPage.value] ?? 0
-  const end = pageBreaks.value[currentPage.value + 1] ?? filteredEndpoints.value.length
-  return filteredEndpoints.value.slice(start, end)
-})
-
-const paginationPages = computed(() =>
-  pageBreaks.value.map((_, i) => ({
-    label: String(i + 1),
-    title: `Page ${i + 1}`,
-    href: '#'
-  }))
-)
-
-const displayedEndpoints = computed(() =>
-  isMeasuring.value ? filteredEndpoints.value : paginatedEndpoints.value
-)
 
 if (isSolution) {
   const topicsAPI = new TopicsAPI({ version: 2 })
@@ -413,36 +231,6 @@ if (isSolution) {
     usefulDataApiFourniesParLaSolution.value = []
   }
 }
-
-const sortedUsefulEndpoints = computed(() => {
-  return usefulDataApiFourniesParLaSolution.value?.slice().sort((a, b) => {
-    const aCustomDescription = customDescriptions.value[a.id]
-    const bCustomDescription = customDescriptions.value[b.id]
-
-    const aOrdre =
-      typeof aCustomDescription?.Ordre === 'number' ? aCustomDescription.Ordre : Infinity
-    const bOrdre =
-      typeof bCustomDescription?.Ordre === 'number' ? bCustomDescription.Ordre : Infinity
-
-    if (aOrdre !== bOrdre) return aOrdre - bOrdre
-
-    const aDescriptionLength =
-      aCustomDescription
-        ?.En_quoi_cette_API_ou_dataset_est_utile_pour_ce_cas_d_usage?.length || 0
-    const bDescriptionLength =
-      bCustomDescription
-        ?.En_quoi_cette_API_ou_dataset_est_utile_pour_ce_cas_d_usage?.length || 0
-    const lengthDiff = bDescriptionLength - aDescriptionLength
-    if (lengthDiff !== 0) return lengthDiff
-
-    return a.fields.Nom.localeCompare(b.fields.Nom)
-  })
-})
-
-watch(filteredEndpoints, () => {
-  currentPage.value = 0
-  if (activeAccordion.value === API_ACCORDION_INDEX) recomputePageBreaks()
-})
 
 // === API/dataset-specific ===
 const apiOrDataset = ref<ApiOrDataset | undefined>(undefined)
@@ -482,14 +270,6 @@ const typeLabel = computed(() => {
   if (t === 'Jeu de données') return 'jeu de données'
   return 'API ou jeu de données'
 })
-
-const datagouvUrlType = (apiOrDataset: ApiOrDataset) => {
-  switch (apiOrDataset.Type) {
-    case 'API': return 'dataservices'
-    case 'Jeu de données': return 'datasets'
-    default: throw new Error(`Unknown type: ${apiOrDataset.Type}`)
-  }
-}
 
 const fetchSolutionsForCategory = async (ids: number[]) => {
   const data = await grist.getRecordsByIds('Solutions', ids)
@@ -568,31 +348,9 @@ const integrationScorePerSolution = computed(() => {
 // Index of the "Par l'API directement" accordion within the shared group (registered first)
 const API_ACCORDION_INDEX = 0
 const activeAccordion = ref(-1)
-
-const ACCORDION_TRANSITION_MS = 350 // DsfrAccordion animation duration
-
-watch(activeAccordion, (val) => {
-  if (val === API_ACCORDION_INDEX) setTimeout(recomputePageBreaks, ACCORDION_TRANSITION_MS)
-})
 </script>
 
 <style scoped>
-
-.fr-table .fr-table__wrapper {
-  border: 1px solid var(--border-default-grey);
-  height: calc(100vh - 13rem);
-  overflow: hidden;
-}
-
-.fr-table .fr-table__wrapper tbody tr {
-  height: var(--row-height, auto);
-}
-
-@media (min-width: 48em) {
-  .fr-table__header .fr-search-bar {
-    min-width: 28rem;
-  }
-}
 
 .reco-card {
   background-color: var(--background-alt-beige-gris-galet);
@@ -605,10 +363,5 @@ watch(activeAccordion, (val) => {
 :deep(.fr-accordion__btn[aria-expanded='true']) {
   background-color: var(--background-alt-blue-france);
   color: var(--text-action-high-blue-france);
-}
-
-.fr-text--sm :deep(p) {
-  font-size: inherit;
-  line-height: inherit;
 }
 </style>
